[Python][Easy][Schema Update] Refactor Test Mock Class
- Relocated Mock Model Parser class to util folder
