+++ python/tests/test_registry.py
@@ -5,30 +5,13 @@ from aiconfig.model_parser import ModelParser
 from aiconfig.registry import ModelParserRegistry
 import pytest
 
+from .util.mock_parser import MockModelParser
+
 
 @pytest.fixture
 def ai_config_runtime():
     runtime = AIConfigRuntime.create("Untitled AIConfig")
     return runtime
-
-
-class MockModelParser(ModelParser):
-    def __init__(self):
-        pass
-
-    def id(self):
-        return "mock_model_parser"
-
-    def serialize(**kwargs):
-        return
-
-    def deserialize(**kwargs):
-        return
-
-    def run(**kwargs):
-        return
-    def get_output_text():
-        return
 
 
 class TestModelParserRegistry:
